\name{abundances}
\alias{abundances}
\docType{data}
\title{
Abundances data
}
\description{
Dummy data set showing abundances species in several locations.
}
\usage{abundances}
\format{
  A data frame (15 x 10) where  each row  contains the abundances of  15  species found in zones E1,..., E10.
}

\examples{
data("abundances")
}
\keyword{datasets}
