\name{microsatt}
\alias{microsatt}
\docType{data}
\title{
Microsatt data
}
\description{
\code{microsatt} includes allele frequencies for 18 cattle breeds (bull or zebu), of French and African descent, typed on 9 microsatellites and they were  compiled from data used in several articles (Moazami-Goudarzi et al. 1997, Souvenir Zafindrajaona et al. 1999, Moazami-Goudarzi et al. 2001). The names of the loci are: INRA5, INRA32, INRA35, INRA63, INRA72, ETH152, ETH225, INRA16 and INRAK
}
\usage{microsatt}
\format{
  A data frame with 18 observations and  112 variables representing allele frequencies.
 }

\source{
Moazami-Goudarzi K., Laloe D. , Furet J. P. , and Grosclaude  F. (1997) Analysis of genetic relationships between 10 cattle breeds with 17 microsatellites.\emph{ Animal Genetics}, \bold{28}, 338--345.

Souvenir Zafindrajaona P.,  Zeuh V.,  Moazami-Goudarzi K.,  Laloe D.,  Bourzat D.,  Idriss A., and  Grosclaude F. (1999) Etude du statut phylognetique du bovin Kouri du lac Tchad a l'aide de marqueurs moleculaires. \emph{Revue d'Elevage et de Medecine Veterinaire des pays Tropicaux}, \bold{55}, 155--162.

Moazami-Goudarzi K.,  Belemsaga D. M. A.,  Ceriotti G., Laloe  D., Fagbohoun F. , Kouagou N. T., Sidibe I., Codjia  V.,  Crimella M. C.,  Grosclaude F., and Toure  S. M. (2001) Caracterisation de la race bovine Somba a l'aide de marqueurs moleculaires. \emph{Revue d'Elevage et de Medecine Veterinaire des pays Tropicaux}, \bold{54}, 1--10.
}

\examples{
\dontrun{
data("microsatt")
names(microsatt)}
}
\keyword{datasets}
